+++ Runtime/UI/Scripts/Vector2UI.cs
@@ -37,10 +37,10 @@ public class Vector2UI : ControllableUI
             RecordUndo();
 
             var list = new List<object>();
-            list.Add(TypeConverter.getFloat(value.ToString()));
-            list.Add(TypeConverter.getFloat(YInput.text.ToString()));
+            list.Add(TypeConverter.GetFloat(value.ToString()));
+            list.Add(TypeConverter.GetFloat(YInput.text.ToString()));
 
-            target.setFieldProp(property, list);
+            target.SetFieldProp(property, list);
         });
 
         YInput.onEndEdit.AddListener((value) =>
@@ -48,10 +48,10 @@ public class Vector2UI : ControllableUI
             RecordUndo();
 
             var list = new List<object>();
-            list.Add(TypeConverter.getFloat(XInput.text.ToString()));
-            list.Add(TypeConverter.getFloat(value.ToString()));
+            list.Add(TypeConverter.GetFloat(XInput.text.ToString()));
+            list.Add(TypeConverter.GetFloat(value.ToString()));
 
-            target.setFieldProp(property, list);
+            target.SetFieldProp(property, list);
         });
     }