+++ Runtime/UI/Scripts/Vector3IntUI.cs
@@ -35,11 +35,11 @@ public class Vector3IntUI : ControllableUI
             RecordUndo();
 
             var list = new List<object>();
-            list.Add(TypeConverter.getInt(value.ToString()));
-            list.Add(TypeConverter.getInt(YInput.text.ToString()));
-            list.Add(TypeConverter.getInt(ZInput.text.ToString()));
+            list.Add(TypeConverter.GetInt(value.ToString()));
+            list.Add(TypeConverter.GetInt(YInput.text.ToString()));
+            list.Add(TypeConverter.GetInt(ZInput.text.ToString()));
 
-            target.setFieldProp(property, list);
+            target.SetFieldProp(property, list);
         });
 
         YInput.onEndEdit.AddListener((value) =>
@@ -47,11 +47,11 @@ public class Vector3IntUI : ControllableUI
             RecordUndo();
 
             var list = new List<object>();
-            list.Add(TypeConverter.getInt(XInput.text.ToString()));
-            list.Add(TypeConverter.getInt(value.ToString()));
-            list.Add(TypeConverter.getInt(ZInput.text.ToString()));
+            list.Add(TypeConverter.GetInt(XInput.text.ToString()));
+            list.Add(TypeConverter.GetInt(value.ToString()));
+            list.Add(TypeConverter.GetInt(ZInput.text.ToString()));
 
-            target.setFieldProp(property, list);
+            target.SetFieldProp(property, list);
         });
 
         ZInput.onEndEdit.AddListener((value) =>
@@ -59,11 +59,11 @@ public class Vector3IntUI : ControllableUI
             RecordUndo();
 
             var list = new List<object>();
-            list.Add(TypeConverter.getInt(XInput.text.ToString()));
-            list.Add(TypeConverter.getInt(YInput.text.ToString()));
-            list.Add(TypeConverter.getInt(value.ToString()));
+            list.Add(TypeConverter.GetInt(XInput.text.ToString()));
+            list.Add(TypeConverter.GetInt(YInput.text.ToString()));
+            list.Add(TypeConverter.GetInt(value.ToString()));
 
-            target.setFieldProp(property, list);
+            target.SetFieldProp(property, list);
         });
     }