+++ Runtime/UI/Scripts/Vector4UI.cs
@@ -49,12 +49,12 @@ public class Vector4UI : ControllableUI
             RecordUndo();
 
             var list = new List<object>();
-            list.Add(TypeConverter.getFloat(value.ToString()));
-            list.Add(TypeConverter.getFloat(YInput.text.ToString()));
-            list.Add(TypeConverter.getFloat(ZInput.text.ToString()));
-            list.Add(TypeConverter.getFloat(WInput.text.ToString()));
+            list.Add(TypeConverter.GetFloat(value.ToString()));
+            list.Add(TypeConverter.GetFloat(YInput.text.ToString()));
+            list.Add(TypeConverter.GetFloat(ZInput.text.ToString()));
+            list.Add(TypeConverter.GetFloat(WInput.text.ToString()));
 
-            target.setFieldProp(property, list);
+            target.SetFieldProp(property, list);
         });
 
         YInput.onEndEdit.AddListener((value) =>
@@ -62,12 +62,12 @@ public class Vector4UI : ControllableUI
             RecordUndo();
 
             var list = new List<object>();
-            list.Add(TypeConverter.getFloat(XInput.text.ToString()));
-            list.Add(TypeConverter.getFloat(value.ToString()));
-            list.Add(TypeConverter.getFloat(ZInput.text.ToString()));
-            list.Add(TypeConverter.getFloat(WInput.text.ToString()));
+            list.Add(TypeConverter.GetFloat(XInput.text.ToString()));
+            list.Add(TypeConverter.GetFloat(value.ToString()));
+            list.Add(TypeConverter.GetFloat(ZInput.text.ToString()));
+            list.Add(TypeConverter.GetFloat(WInput.text.ToString()));
 
-            target.setFieldProp(property, list);
+            target.SetFieldProp(property, list);
         });
 
         ZInput.onEndEdit.AddListener((value) =>
@@ -75,12 +75,12 @@ public class Vector4UI : ControllableUI
             RecordUndo();
 
             var list = new List<object>();
-            list.Add(TypeConverter.getFloat(XInput.text.ToString()));
-            list.Add(TypeConverter.getFloat(YInput.text.ToString()));
-            list.Add(TypeConverter.getFloat(value.ToString()));
-            list.Add(TypeConverter.getFloat(WInput.text.ToString()));
+            list.Add(TypeConverter.GetFloat(XInput.text.ToString()));
+            list.Add(TypeConverter.GetFloat(YInput.text.ToString()));
+            list.Add(TypeConverter.GetFloat(value.ToString()));
+            list.Add(TypeConverter.GetFloat(WInput.text.ToString()));
 
-            target.setFieldProp(property, list);
+            target.SetFieldProp(property, list);
         });
 
         WInput.onEndEdit.AddListener((value) =>
@@ -88,12 +88,12 @@ public class Vector4UI : ControllableUI
             RecordUndo();
 
             var list = new List<object>();
-            list.Add(TypeConverter.getFloat(XInput.text.ToString()));
-            list.Add(TypeConverter.getFloat(YInput.text.ToString()));
-            list.Add(TypeConverter.getFloat(ZInput.text.ToString()));
-            list.Add(TypeConverter.getFloat(value.ToString()));
+            list.Add(TypeConverter.GetFloat(XInput.text.ToString()));
+            list.Add(TypeConverter.GetFloat(YInput.text.ToString()));
+            list.Add(TypeConverter.GetFloat(ZInput.text.ToString()));
+            list.Add(TypeConverter.GetFloat(value.ToString()));
 
-            target.setFieldProp(property, list);
+            target.SetFieldProp(property, list);
         });
     }